Allow fsconsul to parse the updated template syntax used by gosecret.
The latest version of gosecret changed tags from [] to {{ }} and
adopted the go template library. However gosecret no longer generates
these tags so it becomes a burdon to convert gosecret docs to the old
tags for use with fsconsul. This was my least resistance approach to
merging the capabilites of gosercret into the fsconsul watcher. Feel
free to augment to your style.