Bypass decryption when keystore is empty.
No need to bother trying to process templates / decryption if keystore is not available.
Mostly an optimization but also avoids truncating files by calling decrypt with no keystore.
Unknown what the default behavior should be but I prefer having the file delivered with
the goDecrypt tag in place when no keystore is provided.